requirements.txt
Flask==3.0.0
pyserial==3.5
Werkzeug==3.0.0
